Designate main ParameterSentence component for v6
