+++ components/parameter-sentence/index.js
@@ -1,3 +1,3 @@
 export { ParameterSelect as LegacyParameterSelect } from './legacy-parameter-select';
 export { ParameterInputBox as LegacyParameterInputBox } from './legacy-parameter-input';
-export { ParameterSentence as LegacyParameterSentence } from './legacy-parameter-sentence';
+export { ParameterSentence } from './parameter-sentence';
