+++ components/parameter-sentence/legacy-styled.jsx
@@ -62,19 +62,6 @@ export const Container = styled.div`
 	width: ${props => props.width};
 `;
 
-export const ParameterSentence = styled.form.attrs(({ isSearchForm, labelledBy }) => ({
-	role: isSearchForm ? 'search' : 'form',
-	'aria-labelledby': labelledBy,
-}))`
-	/* stylelint-disable no-empty-block https://github.com/stylelint/stylelint/issues/3494 */
-`;
-
-export const Fieldset = styled.fieldset`
-	border: none;
-	padding: 0;
-	margin: 0;
-`;
-
 export const Select = styled.select`
 	appearance: none;
 	user-select: none;
