+++ index-v6.js
@@ -1,4 +1,5 @@
-export { Modal } from './components/modal';
 export { Button, SegmentedButtonGroup } from './components/button';
-export { usePopover, Popover } from './components/popover-v6';
 export { Dropdown } from './components/dropdown';
+export { Modal } from './components/modal';
+export { ParameterSentence } from './components/parameter-sentence';
+export { usePopover, Popover } from './components/popover-v6';
